+++ africanus/averaging/time_and_channel_avg.py
@@ -387,26 +387,116 @@ def chan_normaliser_factory(present):
     return njit(nogil=True, cache=True)(impl)
 
 
-def chan_corr_factory(have_vis, have_flag,
-                      have_weight_spectrum, have_sigma_spectrum):
-    """ Returns function returning number of channels and correlations """
-    if have_vis:
-        def impl(vis, flag, weight_spectrum, sigma_spectrum):
-            return vis.shape[1:]
-    elif have_flag:
-        def impl(vis, flag, weight_spectrum, sigma_spectrum):
-            return flag.shape[1:]
-    elif have_weight_spectrum:
-        def impl(vis, flag, weight_spectrum, sigma_spectrum):
-            return weight_spectrum.shape[1:]
-    elif have_sigma_spectrum:
-        def impl(vis, flag, weight_spectrum, sigma_spectrum):
-            return sigma_spectrum.shape[1:]
-    else:
-        def impl(vis, flag, weight_spectrum, sigma_spectrum):
-            return (1, 1)
+@generated_jit(nopython=True, nogil=True, cache=True)
+def shape_or_invalid_shape(array, ndim):
+    """ Return array shape tuple or (-1,)*ndim if the array is None """
 
-    return njit(nogil=True, cache=True)(impl)
+    try:
+        ndim_lit = getattr(ndim, "literal_value")
+    except AttributeError:
+        raise ValueError("ndim must be a integer literal")
+
+    if is_numba_type_none(array):
+        tup = (-1,)*ndim_lit
+
+        def impl(array, ndim):
+            return tup
+    else:
+        def impl(array, ndim):
+            return array.shape
+
+    return impl
+
+
+@njit(nogil=True, cache=True)
+def find_chan_corr(chan, corr, shape, chan_idx, corr_idx):
+    """
+    1. Get channel and correlation from shape if not set and the shape is valid
+    2. Check they agree if they already agree
+
+    Parameters
+    ----------
+    chan : int
+        Existing channel size
+    corr : int
+        Existing correlation size
+    shape : tuple
+        Array shape tuple
+    chan_idx : int
+        Index of channel dimension in ``shape``.
+    corr_idx : int
+        Index of correlation dimension in ``shape``.
+
+    Returns
+    -------
+    int
+        Modified channel size
+    int
+        Modified correlation size
+    """
+    if chan_idx != -1:
+        array_chan = shape[chan_idx]
+
+        # Corresponds to a None array, ignore
+        if array_chan == -1:
+            pass
+        # chan is not yet set, assign
+        elif chan == 0:
+            chan = array_chan
+        # Check consistency
+        elif chan != array_chan:
+            raise ValueError("Inconsistent Channel Dimension "
+                             "in Input Arrays")
+
+    if corr_idx != -1:
+        array_corr = shape[corr_idx]
+
+        # Corresponds to a None array, ignore
+        if array_corr == -1:
+            pass
+        # corr is not yet set, assign
+        elif corr == 0:
+            corr = array_corr
+        # Check consistency
+        elif corr != array_corr:
+            raise ValueError("Inconsistent Correlation Dimension "
+                             "in Input Arrays")
+
+    return chan, corr
+
+
+@njit(nogil=True, cache=True)
+def chan_corrs(vis, flag,
+               weight_spectrum, sigma_spectrum,
+               chan_freq, chan_width):
+    """
+    Infer channel and correlation size from input dimensions
+
+    Returns
+    -------
+    int
+        channel size
+    int
+        correlation size
+    """
+    vis_shape = shape_or_invalid_shape(vis, 3)
+    flag_shape = shape_or_invalid_shape(flag, 3)
+    weight_spectrum_shape = shape_or_invalid_shape(weight_spectrum, 3)
+    sigma_spectrum_shape = shape_or_invalid_shape(sigma_spectrum, 3)
+    chan_freq_shape = shape_or_invalid_shape(chan_freq, 1)
+    chan_width_shape = shape_or_invalid_shape(chan_width, 1)
+
+    chan = 0
+    corr = 0
+
+    chan, corr = find_chan_corr(chan, corr, vis_shape, 1, 2)
+    chan, corr = find_chan_corr(chan, corr, flag_shape, 1, 2)
+    chan, corr = find_chan_corr(chan, corr, weight_spectrum_shape, 1, 2)
+    chan, corr = find_chan_corr(chan, corr, sigma_spectrum_shape, 1, 2)
+    chan, corr = find_chan_corr(chan, corr, chan_freq_shape, 0, -1)
+    chan, corr = find_chan_corr(chan, corr, chan_width_shape, 0, -1)
+
+    return chan, corr
 
 
 def is_chan_flagged_factory(present):
@@ -476,16 +566,19 @@ def row_chan_average(row_meta, chan_meta, flag_row=None, weight=None,
 
     set_flagged = set_flagged_factory(have_flag)
 
-    chan_corrs = chan_corr_factory(have_vis, have_flag,
-                                   have_weight_spectrum, have_sigma_spectrum)
+    dummy_chan_freq = None
+    dummy_chan_width = None
 
     def impl(row_meta, chan_meta, flag_row=None, weight=None,
              vis=None, flag=None,
              weight_spectrum=None, sigma_spectrum=None):
 
         out_rows = row_meta.time.shape[0]
+        nchan, ncorrs = chan_corrs(vis, flag,
+                                   weight_spectrum, sigma_spectrum,
+                                   dummy_chan_freq, dummy_chan_width)
+
         chan_map, out_chans = chan_meta
-        _, ncorrs = chan_corrs(vis, flag, weight_spectrum, sigma_spectrum)
 
         out_shape = (out_rows, out_chans, ncorrs)
 
@@ -711,15 +804,6 @@ def time_and_channel(time, interval, antenna1, antenna2,
     if not isinstance(chan_bin_size, valid_types):
         raise TypeError("chan_bin_size must be a scalar integer")
 
-    have_vis = not is_numba_type_none(vis)
-    have_flag = not is_numba_type_none(flag)
-    have_weight_spectrum = not is_numba_type_none(weight_spectrum)
-    have_sigma_spectrum = not is_numba_type_none(sigma_spectrum)
-
-    chan_corrs = chan_corr_factory(have_vis, have_flag,
-                                   have_weight_spectrum,
-                                   have_sigma_spectrum)
-
     def impl(time, interval, antenna1, antenna2,
              time_centroid=None, exposure=None, flag_row=None,
              uvw=None, weight=None, sigma=None,
@@ -728,8 +812,9 @@ def time_and_channel(time, interval, antenna1, antenna2,
              weight_spectrum=None, sigma_spectrum=None,
              time_bin_secs=1.0, chan_bin_size=1):
 
-        # Get the number of channels + correlations
-        nchan, ncorr = chan_corrs(vis, flag, weight_spectrum, sigma_spectrum)
+        nchan, ncorrs = chan_corrs(vis, flag,
+                                   weight_spectrum, sigma_spectrum,
+                                   chan_freq, chan_width)
 
         # Merge flag_row and flag arrays
         flag_row = merge_flags(flag_row, flag)